refactor:splash  로딩/초기 화면 분리(#175)
